refactor(tagModal): 피드백 적용
