+++ backend/web/src/frontend/src/views/header/component/CategoryAddModal.vue
@@ -55,5 +55,3 @@ export default {
   },
 };
 </script>
-
-<style scoped></style>
